Refactoring display managers
Creating a common infrastructure to render composed UI (display, edit, update) that
is used for Content items and can be reused for other matters. For instance IDisplay doesn't
have any relationship with Content Management, but just Shapes. IContentDisplay
is a specialization of it for IContent. Creating customized DisplayManager enabled to apply
this level of UI composition to any component, for instance it can be reused to render Content Type
editors (not yet done).