Added error handling to setup. (#584)
* Added error handling to setup.

This fixes an issue where Orchard crashes when it is unable to connect to the database (e.g. due to incorrect connection string).
Whenever an exception occurs during DB initialization, that error is shown on the setup screen.

Fixes #355

* Added error message to Tenant setup screens.

* Removed unnecessary code.

Since the DB connection string is not used or validated during tenant creation, there will never be an DB connection failure due to an invalid connection string. The connection string is only used when actually setting up the tenant.