Tried a version where without precompiled fluid page, just provide a little razor content.
Pros: the "razor" page is created and activated as with other razor views.
Pros: Harmonize the way templates from shared files / tenant database are implemented.
Cons: add a little razor compile time but it is fast.
Pros: Simplify the code in some places.
The razor contents are provided at a lower level and for all fluid file providers
So, the "Orchard.Templates" custom fileProvider is simpler.