Prevent unnecessary compilation in incremental build (#1409)
There are custom targets that generate modules.*.map files
to be embedded as resources. These files were regenerated
on every incremental build causing the CoreCompile target
to always see at least one input file that was newer than
the output assembly. This caused the assemblies to be
recompiled on every command-line build. (In VS, the "fast
up-to-date" check masks the problem.)