Reverting making WF definition part of WF instance
I initially thought this was a clever idea because it prevents issues when changing an WF definition while there are halted WF instances, but after further analysis concluded that it was unreasonably inefficient in terms of data size. Especially if we want the engine to be able to perform well when many instances execute. Plus when keeping execution history around, that wouldn't be very efficient.