Add UsersMustValidateEmail in RegistrationSettings
If enabled, set user.IsEmailConfirmed to false and send a ConfirmEmail to validate the user email on Register.
If not enabled, do as if the user IsEmailConfirmed is true on Register.
If enabled, check if email is confirmed on Login.

In Users admin, allow to check IsEmailConfirmed when you edit a user.