fixed the js events binding and the issue related to multiple media fields in one contenttype
attachedMediaField.js
Removed the binding of the drop and dragover javascript events to avoid the issue that disable the moving of the selected text and image moving with mouse in the html editor's.
Fixed the code to generate a javascript alert if user try to upload unsupported file in the media field.
Fixed code to remove item from the mediaItems structure to uniform the behavior between the two types of mediafield, to show the card "no image" if there is no element in the field.

Use field ID as a prefix, to create ID for DOM elements unique.

Added shapes to avoid multiple loading of the same items (localization items and scripts).

Changed fields template to improve the bootstrap 4 compatibilities.