restored functionality defined in the issue #4985 and applied some fixes to other AdminMenu's links
restored the new button specific for each contentitems when user filter by contenttype or the user access to contentitem list interface by link for a single contenttype in the AdminMenu.

fixed behavior of the Admin Menu's links Menus and Content Items: if they are not defined inside a recipe (like in the BlogTheme), they have a dynamic behavior and it cause issue if user define link for specific contenttype inside the AdminMenu.