Update SiteSettings.cs
I think it should be set to false, because in some cases (such as internal networks), the CDN server is not always available.
For example, I am from China. My network problem directly caused me to be unable to access some key js file resources, which also caused me to be unable to click the login button. Even if the login is successful, I cannot expand the settings menu.