Cleanup TFM commented node and unneeded whitespace
In the csproj files, remove the unnecessary <TargetFramework> node.  It
was previously commented out due to being set in the AspNetCore
dependencies props file in the Orchard.Build directory.

Since that file eventually gets included in all project builds due to
Directory.Build.props inheritance down the directory subtree, setting
the <TargetFramework> on each node is redundant.  In addition, it adds
additional complexity when multi-targetting the build for more than one
target framework.

With the change the enable multi-targetting for netcoreapp3.1 and
net5.0, it became simpler to remove the node from the individual project
files altogether and control it at a centralized place.

This can be overridden if a specific project needs to target a single
framework but at this time, there is no known need to do so.