Default to net5.0 and update AspNetCoreVersion
This adds back in a default AspNetCoreTargetFramework var for use in
projects that need to execute directly. In addition, both the package
versions for net5.0 and netcoreapp3.1 asp.net core packages were bumped.

Lastly, packages that have issues that need to be addressed for net5.0
are handled in a subsequent ItemGroup with comments about why the
package version needs to deviate from the common version.  In several
cases, this is due to the package not having a release of that version.
In another case, the newest package is causing new warnings for obsolete
code.  Rather than make functional changes during the upgrade to net5.0
multi-targeting, this can be addressed in subsequent focused on updating
that feature.