Default web and console apps to net5.0 TFM
Rather than specify the target framework when running
OrchardCore.Cms.Web, this change defaults it to use net5.0.

Several other projects are also similarly changed.

One wrinkle introduced with this is that the OrchardCore.Tests.csproj
references the OrchardCore.Cms.Web.csproj project via a
<ProjectReference> MSBuild item.  Since the OrchardCore.Tests project
needs to multi-target in order to run tests for both frameworks, the
project reference to OrchardCore.Cms.Web failed when
OrchardCore.Cms.Web was defaulted to use the net5.0 TFM.  To fix this,
the SetTargetFramework attribute is used in the OrchardCore.Tests.csproj
when referencing OrchardCore.Cms.Web.csproj.