Update OrchardCore.Title Doc
In addition to a missing word, "to," I noticed in general the Orchard Core docs have lots of single sentences with single newlines where it really should be either (1) all part of one paragraph, or (2) two newlines marking a new paragraph.  If you look at the first paragraph of this doc before my edit you have two sentences with a line break (but not a full paragraph space) which isn't really normal English writing style...  I'm not sure how much of this type of writing style suggestions you might be interested in getting but I find it a bit awkward to read as-is.  I'm happy to help a bit as I'm reading your docs but only want to do this type of edit if it is welcome...  please just let me know.