Fix OC version used in templates (#12757)
The template default settings are generated on build. These templates have an `OrchardVersion` argument that has a default value defined in an msbuild script to be replaced with the `Version` property. In the release pipeline this value will have the `VersionSuffix` applied to it, like `preview`. 

This PR forces the value of the version property to be the one in the tag which will prevent `preview` from being added. This will also fix the templates' default value for `OrchardVersion`.

The `pack` command uses the `Version` property only for the packages, and at that point the templates json files are already built with the wrong version.