Remove npm assets from dependabot
Based on the following query we close all the PRs it's creating for npm: https://github.com/OrchardCMS/OrchardCore/pulls?q=is%3Apr+author%3Aapp%2Fdependabot+is%3Aclosed

I think it could work out but we should add some functional test of the groups it's updating. And maybe group the dependencies for each feature we test. e.g., graphiql