Remove esproj
Was causing issues with Nuget restore and also was automatically executing npm install when cleaning solution and restoring dependencies. We should not use `npm install` but `yarn install` else this will create a package-lock.json file in each folders for nothing. The only thing we need is the main yarn.lock file from the root folder.